middleware set up for project, post request working
