+++ api/server.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const helmet = require("helmet");
+const {logger} = require("../middleware/middleware");
 
 const server = express();
 const projectRouter = require("../routes/projectRouter");
@@ -7,6 +8,7 @@ const actionRouter = require("../routes/actionRouter");
 
 server.use(helmet());
 server.use(express.json());
+server.use(logger);
 server.use("/api/projects", projectRouter);
 server.use("/api/actions", actionRouter);
